+++ spec/teamtailor/parser_spec.rb
@@ -158,6 +158,19 @@ RSpec.describe Teamtailor::Parser do
     end
   end
 
+  context 'parsing locations' do
+    it 'works' do
+      payload = File.read 'spec/fixtures/v1/locations.json'
+      json_payload = JSON.parse payload
+
+      result = Teamtailor::Parser.parse json_payload
+
+      expect(result.size).to eq 1
+      expect(result.map(&:id)).to include(1)
+      expect(result.map(&:name)).to include('Stockholm')
+    end
+  end
+
   context 'getting an unknown record' do
     it 'raises an Teamtailor::UnknownResponseTypeError' do
       payload = { 'id' => 3, 'type' => 'foo' }